+++ src/step02_get_pubmed.py
@@ -1,5 +1,6 @@
-
-from triplea.service.repository.state.initial import get_article_list_from_pubmed_all_store_to_arepo
+from triplea.service.repository.state.initial import (
+    get_article_list_from_pubmed_all_store_to_arepo,
+)
 
 if __name__ == "__main__":
     # Get article from Pubmed