+++ CHANGELOG.md
@@ -3,10 +3,11 @@ All notable changes to this project will be documented in this file.
 
 ## v0.0.1 2024-03-24
 
-### Task
-
 
 ### Improvements
+- Code Quality with flake8
+- Improve Documentation
+- Add `CITATION.cff` 2024-09-24
 - Add & Improve `step08_generate_co_occurrence_graph`
 - Change structure And step
 - Add `step11_topic_extract_with_llm`